brcmsmac: AP mode: update beacon when TIM changes
Beacons+TIM are created/updated for beaconing only when BSS_CHANGED_BEACON.
This is not compliant with power-saving client stations as the beacons do not
have valid TIM and can cause the network to stall at random occasions with
highly variable latencies.
Fix it by updating beacon templates on mac80211 set_tim callback.
Addresses an issue described in:
https://marc.info/?i=20180911163534.21312d08%20()%20manjaro